+++ generators/view/writing/htmlFiles.js
@@ -13,7 +13,6 @@
 	const utils = require('../../common/utils.js');
 
 	module.exports = $that => {
-
 		const filePrefix = 'app/views/' + $that.viewNameCamel + '/' + $that.viewNameCamel + '.';
 		const config     = {
 			viewNameCamel: $that.viewNameCamel,